virt_vm:Check if log file exists before unlink it

+++ virttest/virt_vm.py
@@ -1146,7 +1146,8 @@ class BaseVM(object):
             )
         except Exception:
             utils_logfile.close_log_file(log_filename)
-            os.unlink(log_filename)
+            if os.path.exists(log_filename):
+                os.unlink(log_filename)
             raise
         session.close_hooks += [utils_logfile.close_own_log_file(log_filename)]
         session.set_status_test_command(self.params.get("status_test_command", ""))
